requirements.txt
biopython==1.85
contourpy==1.3.3
cycler==0.12.1
fonttools==4.59.0
kiwisolver==1.4.8
markdown-it-py==3.0.0
matplotlib==3.10.3
mdurl==0.1.2
numpy==2.3.2
packaging==25.0
pillow==11.3.0
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
rich==14.1.0
six==1.17.0
